+++ src/m_misc.c
@@ -28,6 +28,7 @@ static const char
 rcsid[] = "$Id: m_misc.c,v 1.6 1997/02/03 22:45:10 b1 Exp $";
 
 #include <stdlib.h>
+#include <stdint.h>
 #include <ctype.h>
 
 extern int access(char *file, int mode);
@@ -311,7 +312,7 @@ extern byte	scantokey[128];
 void M_LoadDefaults (void)
 {
     int		i;
-    int		len;
+    int64_t	len;
     FILE*	f;
     char	def[80];
     char	strparm[100];
@@ -362,8 +363,7 @@ void M_LoadDefaults (void)
 			if (!isstring)
 			    *defaults[i].location = parm;
 			else
-			    *defaults[i].location =
-				(int) newstring;
+			    *defaults[i].location = newstring;
 			break;
 		    }
 	    }
